Added pageInfo Pill
Added pageInfo option. The pageInfo option will display a pill at the bottom of the view with the current page number.

Other options added are: 
pageInfo: bool (enable the control, default is false),
pageInfoBackgroundColor: string (set the backgroundColor of the pill, default is black with 0.25 alpha),
pageInfoTextStyle: text style (set the text style of the pill, no default set),
pageInfoTextSeparator: string (set the character used to separate the current page from the amount of pages eg: ' of ' would show 1 of 3. Default is ' / ')