Check children's keys before resetting currentPage.
This allows the children to still be dynamic and reset the carousel
properly, but does so by checking for key equality and position, rather
than component equality.

Setting keys is a common practice for an array of React elements, so
those use cases in which an array of elements is being rendered will fit
in nicely with this scheme. Those that don't use keys will not have the
resetting behavior.

I'll leave documenting the use of keys for children to a separate
commit.